requirements.txt
accelerate==0.24.1
aider-chat==0.17.0
aiohttp==3.8.6
aiosignal==1.3.1
altair==5.1.2
annotated-types==0.6.0
anyio==3.7.1
APScheduler==3.6.3
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
blinker==1.7.0
cachetools==4.2.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.0
ConfigArgParse==1.7
debugpy==1.8.0
decorator==5.1.1
diskcache==5.6.3
distro==1.8.0
exceptiongroup==1.1.3
executing==2.0.1
filelock==3.13.1
FLAML==2.1.1
flatbuffers==23.5.26
frozenlist==1.4.0
fsspec==2023.10.0
gitdb==4.0.11
GitPython==3.1.40
grep-ast==0.2.4
h11==0.14.0
httpcore==1.0.1
httpx==0.25.1
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
ipykernel==6.26.0
ipython==8.17.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter-client==8.6.0
jupyter-core==5.5.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
nest-asyncio==1.5.8
networkx==3.2.1
numpy==1.26.1
onnxruntime==1.16.2
openai==0.28.1
packaging==23.2
pandas==2.1.2
parso==0.8.3
pathspec==0.11.2
Pillow==10.1.0
platformdirs==4.0.0
prompt-toolkit==3.0.39
protobuf==4.25.0
psutil==5.9.6
psycopg2-binary
pure-eval==0.2.2
pyarrow==14.0.1
pyautogen==0.1.14
pycparser==2.21
pydantic==2.4.2
pydantic-core==2.10.1
pydeck==0.8.1b0
Pygments==2.16.1
pyreadline3==3.4.1
python-crfsuite==0.9.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==13.7
pytz==2023.3.post1
pyvi==0.1.1
regex==2023.10.3
replicate==0.17.0
requests==2.31.0
rich==13.6.0
rpds-py==0.10.6
safetensors==0.4.0
scikit-learn==1.3.2
scipy==1.11.3
six==1.16.0
sklearn-crfsuite==0.3.6
smmap==5.0.1
sniffio==1.3.0
sounddevice==0.4.6
soundfile==0.12.1
stack-data==0.6.3
streamlit==1.28.1
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.3.0
threadpoolctl==3.2.0
tiktoken==0.5.1
tokenizers==0.14.1
toml==0.10.2
toolz==0.12.0
torch==2.1.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
transformers==4.35.0
tree-sitter==0.20.2
tree-sitter-languages==1.8.0
typing-extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
Unidecode==1.3.7
urllib3==2.0.7
validators==0.22.0
watchdog==3.0.0
wcwidth==0.2.9
yarl==1.9.2
zipp==3.17.0
